requirements-debug.txt
-r requirements.txt

# Required for tests
Faker==8.12.0
factory_boy==3.2.0

djangorestframework-stubs~=1.3.0
django_nplusone
django-admin-generator